requirements.txt
streamlit==1.42.0
supabase==2.12.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
